+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Adds Stimulus generator. ([@abhaynikam][])
 * Adds Rails Admin generator. ([@abhaynikam][])
 * Adds Paper Trail generator. ([@abhaynikam][])
+* Update default node-version for GitHub Action generator and fixes PG setup issues. ([@abhaynikam][])
 
 ## 0.10.0 (May 26th, 2021)
 * Updates Stripe payment generator. ([@abhaynikam][])